Split bundle builders
