+++ bee-bundle/src/lib.rs
@@ -3,14 +3,14 @@ extern crate serde;
 
 mod bundle;
 mod constants;
+mod incoming_bundle_builder;
+mod outgoing_bundle_builder;
 mod transaction;
 
-// TODO rename crate to model ?
-pub use crate::bundle::{
-    IncomingBundleBuilder, IncomingBundleBuilderError, OutgoingBundleBuilder,
-    OutgoingBundleBuilderError,
-};
+pub use crate::bundle::Bundle;
+pub use incoming_bundle_builder::{IncomingBundleBuilder, IncomingBundleBuilderError};
+pub use outgoing_bundle_builder::{OutgoingBundleBuilder, OutgoingBundleBuilderError};
 pub use transaction::{
-    Address, Hash, Index, Nonce, Payload, Tag, Timestamp, Transaction, TransactionBuilder,
-    TransactionBuilderError, TransactionBuilders, Transactions, Value,
+    Address, Hash, Index, Nonce, Payload, Tag, Timestamp, Transaction, TransactionBuilder, TransactionBuilderError,
+    TransactionBuilders, Transactions, Value,
 };
